Created benchmark to compare metric implementation
